requirements.txt
# vLLM 推理与多 LoRA 服务
vllm>=0.6.0

# ModelScope 下载专家模型
modelscope>=1.11.0

# 可选：OpenAI 兼容客户端示例
openai>=1.0.0

# 内网 gRPC 服务（The Internal Server）
grpcio>=1.60.0
grpcio-tools>=1.60.0

# 对外 Web 网关（协议转换）
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
